dont think we really need long here - we will be in trouble if people are adding that many elements to their models :-)

+++ src/sbml/validator/constraints/LocalParameterMathCheck.cpp
@@ -141,17 +141,17 @@ LocalParameterMathCheck::checkCiElement (const Model& m,
   {
     if (m.getLevel() >= 3)
     {
-      for (unsigned long rxn = 0; rxn < m.getNumReactions(); rxn++)
+      for (unsigned int rxn = 0; rxn < m.getNumReactions(); rxn++)
       {
         const Reaction* reaction = m.getReaction(rxn);
-        for (unsigned long reactant = 0; reactant < reaction->getNumReactants(); reactant++)
+        for (unsigned int reactant = 0; reactant < reaction->getNumReactants(); reactant++)
         {
           if (reaction->getReactant(reactant)->getIdAttribute() == name)
           {
             return;
           }
         }
-        for (unsigned long product = 0; product < reaction->getNumProducts(); product++)
+        for (unsigned int product = 0; product < reaction->getNumProducts(); product++)
         {
           if (reaction->getProduct(product)->getIdAttribute() == name)
           {
